# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LiveKit Meet API",
    "version": "1.0.0",
    "description": "REST API for managing LiveKit-powered meetings, participants, tokens, chat history, and scheduler operations."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "schemas": {
      "Meeting": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "startTime": { "type": "string", "format": "date-time" },
          "endTime": { "type": "string", "format": "date-time" },
          "status": { "type": "string", "enum": ["SCHEDULED", "ACTIVE", "ENDED"] },
          "isPublic": { "type": "boolean" },
          "isInstant": { "type": "boolean" },
          "roomName": { "type": "string" },
          "joinUrl": { "type": "string", "format": "uri" },
          "participants": {
            "type": "array",
            "items": { "type": "string", "format": "email" }
          }
        },
        "required": [
          "id",
          "title",
          "startTime",
          "endTime",
          "status",
          "isPublic",
          "isInstant",
          "roomName",
          "joinUrl",
          "participants"
        ]
      },
      "MeetingCreateRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "startTime": { "type": "string", "format": "date-time" },
          "endTime": { "type": "string", "format": "date-time" },
          "isPublic": { "type": "boolean", "default": false },
          "password": { "type": "string", "minLength": 6 },
          "participants": {
            "type": "array",
            "items": { "type": "string", "format": "email" }
          }
        },
        "required": ["title", "startTime", "endTime", "password"]
      },
      "InstantMeetingRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "isPublic": { "type": "boolean", "default": false },
          "participants": {
            "type": "array",
            "items": { "type": "string", "format": "email" }
          }
        },
        "required": ["title"]
      },
      "MeetingListResponse": {
        "type": "object",
        "properties": {
          "meetings": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Meeting" }
          }
        },
        "required": ["meetings"]
      },
      "MeetingCreateResponse": {
        "type": "object",
        "properties": {
          "meeting": { "$ref": "#/components/schemas/Meeting" },
          "password": { "type": "string" }
        },
        "required": ["meeting", "password"]
      },
      "TokenRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" },
          "name": { "type": "string" },
          "metadata": { "type": "string" }
        },
        "required": ["email", "password"]
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "livekitUrl": { "type": "string", "format": "uri" },
          "roomName": { "type": "string" },
          "meeting": { "$ref": "#/components/schemas/Meeting" }
        },
        "required": ["token", "livekitUrl", "roomName", "meeting"]
      },
      "ParticipantsUpdateRequest": {
        "type": "object",
        "properties": {
          "participants": {
            "type": "array",
            "items": { "type": "string", "format": "email" }
          }
        },
        "required": ["participants"]
      },
      "Message": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "meetingId": { "type": "string", "format": "uuid" },
          "senderEmail": { "type": "string", "format": "email", "nullable": true },
          "senderName": { "type": "string", "nullable": true },
          "content": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "meetingId", "content", "createdAt"]
      },
      "MessageCreateRequest": {
        "type": "object",
        "properties": {
          "meetingId": { "type": "string", "format": "uuid" },
          "content": { "type": "string" },
          "senderEmail": { "type": "string", "format": "email" },
          "senderName": { "type": "string" }
        },
        "required": ["meetingId", "content"]
      },
      "MessageListResponse": {
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Message" }
          }
        },
        "required": ["messages"]
      },
      "SchedulerResponse": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string", "format": "date-time" },
          "activated": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Meeting" }
          },
          "ended": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Meeting" }
          }
        },
        "required": ["timestamp", "activated", "ended"]
      }
    },
    "parameters": {
      "MeetingId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" },
        "description": "Unique identifier of the meeting"
      }
    }
  },
  "paths": {
    "/api/meetings": {
      "get": {
        "summary": "List meetings",
        "description": "Retrieve recent meetings. Filter by room name, status, or limit results.",
        "parameters": [
          {
            "name": "roomName",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string", "enum": ["SCHEDULED", "ACTIVE", "ENDED"] }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 200 }
          }
        ],
        "responses": {
          "200": {
            "description": "List of meetings",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MeetingListResponse" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create scheduled meeting",
        "description": "Create a scheduled meeting with password protection and optional participant allowlist.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MeetingCreateRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Meeting created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MeetingCreateResponse" }
              }
            }
          },
          "400": { "description": "Validation error" }
        }
      }
    },
    "/api/meetings/instant": {
      "post": {
        "summary": "Create instant meeting",
        "description": "Launch an immediate meeting with a generated password and join link.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/InstantMeetingRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Instant meeting created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MeetingCreateResponse" }
              }
            }
          },
          "400": { "description": "Validation error" }
        }
      }
    },
    "/api/meetings/{id}/token": {
      "post": {
        "summary": "Issue LiveKit access token",
        "description": "Validate join credentials and issue a LiveKit token for the meeting.",
        "parameters": [
          { "$ref": "#/components/parameters/MeetingId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TokenRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token issued",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenResponse" }
              }
            }
          },
          "401": { "description": "Invalid password" },
          "403": { "description": "Participant not authorized" },
          "404": { "description": "Meeting not found" }
        }
      }
    },
    "/api/meetings/{id}/participants": {
      "get": {
        "summary": "List participants",
        "parameters": [
          { "$ref": "#/components/parameters/MeetingId" }
        ],
        "responses": {
          "200": {
            "description": "Participant emails",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "participants": {
                      "type": "array",
                      "items": { "type": "string", "format": "email" }
                    }
                  },
                  "required": ["participants"]
                }
              }
            }
          },
          "404": { "description": "Meeting not found" }
        }
      },
      "post": {
        "summary": "Replace participant allowlist",
        "parameters": [
          { "$ref": "#/components/parameters/MeetingId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ParticipantsUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated participant list" },
          "400": { "description": "Validation error" },
          "404": { "description": "Meeting not found" }
        }
      },
      "delete": {
        "summary": "Remove participant from allowlist",
        "parameters": [
          { "$ref": "#/components/parameters/MeetingId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" }
                },
                "required": ["email"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated participant list" },
          "400": { "description": "Cannot delete last participant" },
          "404": { "description": "Meeting or participant not found" }
        }
      }
    },
    "/api/scheduler": {
      "get": {
        "summary": "Run scheduler cycle (GET)",
        "description": "Trigger meeting lifecycle updates and obtain a summary.",
        "responses": {
          "200": {
            "description": "Scheduler results",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SchedulerResponse" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Run scheduler cycle (POST)",
        "description": "Trigger meeting lifecycle updates when invoked by cron or background jobs.",
        "responses": {
          "200": {
            "description": "Scheduler results",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SchedulerResponse" }
              }
            }
          }
        }
      }
    },
    "/api/messages": {
      "get": {
        "summary": "List chat messages",
        "parameters": [
          {
            "name": "meetingId",
            "in": "query",
            "schema": { "type": "string", "format": "uuid" },
            "required": true
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 500 }
          }
        ],
        "responses": {
          "200": {
            "description": "Meeting chat history",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageListResponse" }
              }
            }
          },
          "400": { "description": "Missing meetingId" }
        }
      },
      "post": {
        "summary": "Persist chat message",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MessageCreateRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Chat message stored",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Message" }
              }
            }
          },
          "404": { "description": "Meeting not found" }
        }
      }
    }
  }
}